Fix data type for data_limit and expire_duration in user_templates Added multiMode for gRPC network Bug fix

+++ app/db/models.py
@@ -106,8 +106,8 @@ class UserTemplate(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(64), nullable=False, unique=True)
-    data_limit = Column(Integer, default=0)
-    expire_duration = Column(Integer, default=0)  # in seconds
+    data_limit = Column(BigInteger, default=0)
+    expire_duration = Column(BigInteger, default=0)  # in seconds
     username_prefix = Column(String(20), nullable=True)
     username_suffix = Column(String(20), nullable=True)
